backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.6
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
openai>=1.0.0
scikit-learn>=1.3.0
celery>=5.3.0
redis>=4.6.0
python-dotenv>=1.0.0
